Code optimization: 1 Adjust the code position 2 Add a unit test file
