+++ client/lcd_test.go
@@ -1,0 +1,20 @@
+package client
+
+import (
+	"github.com/irisnet/irishub/app"
+	"github.com/spf13/viper"
+	"testing"
+)
+
+func TestRestServer(t *testing.T) {
+	cdc := app.MakeCodec()
+	comm := ServeCommand(cdc)
+	viper.Set("chain-id", "fuxi-1001")
+	viper.Set("node", "tcp://localhost:26657")
+	viper.Set("laddr", "tcp://localhost:1317")
+
+	comm.ExecuteC()
+}
+
+
+
