+++ client/tx.go
@@ -1,4 +1,4 @@
-package main
+package client
 
 import (
 	"encoding/json"
